# front-cinepro/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - CinePro",
    "description": "CinePro API Documentation",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "N/A",
      "url": "https://to_be_determinated.com/",
      "email": "to_be_determinated@coding.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local ENV"
    },
    {
      "url": "https://to_be_determinated.com",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/films/update/{filmId}": {
      "put": {
        "tags": ["film-controller"],
        "operationId": "updateFilm",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "titre",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "langue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "soustitre",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doublage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "duree",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "videoUrl",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "listeActeurs",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "listeRealisateurs",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "titreOriginal",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categorie",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateDeSortie",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/films/ajouter": {
      "post": {
        "tags": ["film-controller"],
        "operationId": "uploadFilm",
        "parameters": [
          {
            "name": "titre",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "langue",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "soustitre",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doublage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "titreOriginal",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categorie",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "listeActeurs",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "listeRealisateurs",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "duree",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "videoUrl",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateDeSortie",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/films/seances/{filmId}": {
      "get": {
        "tags": ["film-controller"],
        "operationId": "getFilmSeances",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Seance"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/films/image/{filmId}": {
      "get": {
        "tags": ["film-controller"],
        "operationId": "getFilmImage",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          }
        }
      }
    },
    "/films/film/{id}": {
      "get": {
        "tags": ["film-controller"],
        "operationId": "getFilm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Film"
                }
              }
            }
          }
        }
      }
    },
    "/films/cinemas/{filmId}": {
      "get": {
        "tags": ["film-controller"],
        "operationId": "getFilmCinemas",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cinema"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cinemas/{id}": {
      "get": {
        "tags": ["cinema-controller"],
        "operationId": "getCinemaById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Cinema"
                }
              }
            }
          }
        }
      }
    },
...
    "/films/delete/{filmId}": {
      "delete": {
        "tags": ["film-controller"],
        "operationId": "deleteFilm",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cinema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nomCinema": {
            "type": "string"
          },
          "nbreDeSalles": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Film": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "langue": {
            "type": "string"
          },
          "soustitre": {
            "type": "string"
          },
          "doublage": {
            "type": "string"
          },
          "titre": {
            "type": "string"
          },
          "titreOriginal": {
            "type": "string"
          },
          "categorie": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "duree": {
            "type": "string"
          },
          "videoUrl": {
            "type": "string"
          },
          "listeActeurs": {
            "type": "string"
          },
          "listeRealisateurs": {
            "type": "string"
          },
          "dateDeSortie": {
            "type": "string",
            "format": "date-time"
          },
          "classement": {
            "type": "integer",
            "format": "int32"
          },
          "seances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Seance"
            }
          }
        }
      },
      "Seance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "heureDebut": {
            "type": "string",
            "format": "date-time"
          },
          "ultraAVX": {
            "type": "boolean"
          },
          "imax": {
            "type": "boolean"
          },
          "troisD": {
            "type": "boolean"
          },
          "sieges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Siege"
            }
          }
        }
      },
      "Siege": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "numeroSiege": {
            "type": "string"
          },
          "rangeeSiege": {
            "type": "string"
          },
          "sectionSiege": {
            "type": "string"
          },
          "estReserve": {
            "type": "boolean"
          },
          "seance": {
            "$ref": "#/components/schemas/Seance"
          }
        }
      },
      "Image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "imageId": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}